requirements.txt
streamlit
pandas
numpy
pandasql
matplotlib
langchain-core
langchain-openai
langchain-community
chromadb
python-dotenv